Extract `lib.user.info`
This also removes the unused `has_positive_balance` function (unused
as a template filter)
